requirements.txt
numpy==1.23.4
onnx==1.14.0
onnxruntime==1.15.1
opencv_python==4.7.0.72
opencv_python_headless==4.6.0.66
Pillow==9.3.0
Pillow==10.0.0
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.64.1
